--- src/app/components/BackgroundGraphic/style.module.scss
@@ -1,14 +1,0 @@
-.backgroundImage {
-  position: absolute;
-  margin: 0;
-  height: 100%;
-  z-index: 0;
-
-  &.left {
-    left: 0;
-  }
-
-  &.right {
-    right: 0;
-  }
-}
