+++ src/app/globals.css
@@ -6,9 +6,8 @@
 }
 
 html,
-body {
-  width: 100%;
-  height: 100%;
+body,
+footer {
   margin: 0;
   padding: 0;
   background: linear-gradient(180deg, #a9e0fff2 25.52%, #ffe2b6 76.79%, #ffacac 98.6%);
