+++ src/app/page.module.scss
@@ -7,7 +7,7 @@
 
   margin: 0 var(--side-padding);
 
-  @media screen and (max-width: 600px) {
+  @media screen and (max-width: 768px) {
     margin: 0 var(--side-padding-small);
   }
