+++ src/components/background/style.module.scss
@@ -1,0 +1,28 @@
+.backgroundImage {
+  position: absolute;
+  margin: 0;
+  height: 100%;
+  z-index: 0;
+
+  &.left {
+    left: 0;
+  }
+
+  &.right {
+    right: 0;
+  }
+
+  &.mobile {
+    display: none;
+  }
+
+  @media screen and (max-width: 768px) {
+    &.mobile {
+      display: block;
+    }
+
+    &.desktop {
+      display: none;
+    }
+  }
+}
