+++ src/sections/Footer/index.tsx
@@ -1,5 +1,5 @@
-import Socials from '@/components/Socials';
-import styles from './footer.module.scss';
+import Socials from '@/components/socials';
+import styles from './style.module.scss';
 import Image from 'next/image';
 
 export default function Footer() {
@@ -9,10 +9,9 @@ export default function Footer() {
         <Image src="/acm-logo.png" alt="ACM Logo" width={100} height={100} />
 
         <h1 className={styles.title}>
-          <span className={styles.acm}>ACM</span> at UCSD
+          <span className={styles.acm}>ACM</span> at UCSD x TESC
         </h1>
       </a>
-
       <Socials />
     </div>
   );
