+++ src/sections/Footer/style.module.scss
@@ -1,8 +1,7 @@
 .container {
   display: flex;
   justify-content: space-between;
-
-  width: 100%;
+  margin: 7.5rem;
 }
 
 .left {
@@ -10,14 +9,13 @@
   align-items: center;
   gap: 1rem;
 
-  color: inherit;
+  color: black;
   text-decoration: none;
 }
 
 .title {
   font-size: 32px;
   font-weight: normal;
-  font-family: inherit;
 }
 
 .acm {
@@ -35,5 +33,6 @@
     flex-direction: column;
     align-items: center;
     gap: 0;
+    text-align: center;
   }
 }
