+++ src/sections/landing/hero/style.module.scss
@@ -4,6 +4,10 @@
   display: flex;
   align-items: center;
   justify-self: start;
+
+  @media screen and (max-width: 768px) {
+    align-items: flex-start;
+  }
 }
 
 .cta {
@@ -26,6 +30,10 @@
     font-style: normal;
     font-weight: 500;
     line-height: 32px;
+  }
+
+  @media screen and (max-width: 768px) {
+    margin-top: 25vh;
   }
 
   @media (max-width: 600px) {
