+++ src/sections/landing/tracks/style.module.scss
@@ -1,6 +1,6 @@
 .tracks {
   width: 100%;
-  min-height: 100vh;
+  height: fit-content;
 
   h2 {
     font-size: 36px;
