+++ src/sections/landing/welcome/index.tsx
